+++ tests/fixtures/issue-2377-whole-object-use/src/attribute.tsx
@@ -1,0 +1,9 @@
+import * as AttributeIcons from './attribute-icons';
+import { Callout } from './callout';
+
+export const Attribute = () => (
+  <div>
+    <AttributeIcons.AttributeStar />
+    <Callout icons={AttributeIcons} />
+  </div>
+);